requirements.txt
django==2.1.2
pytz==2018.5
djangorestframework==3.8.2
shortuuid==0.5.0
numpy==1.15.2
aspy.yaml==1.1.1
cached-property==1.5.1
cfgv==1.1.0
identify==1.1.7
nodeenv==1.3.2
pre-commit==1.11.1
pyyaml==3.13
six==1.11.0
toml==0.10.0
virtualenv==16.0.0
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==2.4.0
django-extensions==2.1.3
mysqlclient==1.3.12